+++ src/main/java/ut/veikotiit/checkers/scorer/BitBoardScorer.java
@@ -3,5 +3,5 @@ package ut.veikotiit.checkers.scorer;
 import ut.veikotiit.checkers.bitboard.BitBoard;
 
 public interface BitBoardScorer {
-    int getScore(BitBoard bitBoard);
+    double getScore(BitBoard bitBoard);
 }
